backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
dj-database-url==2.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2

